requirements.txt
streamlit
pandas
cohere
requests
matplotlib
python-dotenv
cachetools